% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_lisi}
\alias{compute_lisi}
\title{Compute Local Inverse Simpson's Index (LISI)}
\usage{
compute_lisi(X, meta_data, label_colnames, perplexity = 30, nn_eps = 0)
}
\arguments{
\item{X}{A matrix with cells (rows) and features (columns).}

\item{meta_data}{A data frame with one row per cell.}

\item{label_colnames}{Which variables to compute LISI for.}

\item{perplexity}{The effective number of each cell's neighbors.}

\item{nn_eps}{Error bound for nearest neighbor search with \code{RANN:nn2()}.
Default of 0.0 implies exact nearest neighbor search.}
}
\value{
A data frame of LISI values. Each row is a cell and each
column is a different label variable.
}
\description{
Use this function to compute LISI scores of one or more labels.
NOTE: Forked from https://github.com/immunogenomics/LISI
}
\examples{

## Example with 400 cells. 
library(scIntegrationMetrics)
library(dplyr)
library(tidyr)
library(ggplot2)
library(magrittr)

head(scIntegrationMetrics::meta_data)

## Let's color cells by labels. For label 1, there are mixed and non-mixed
## groups. For label 2, all cells are well mixed. 
scIntegrationMetrics::X \%>\% 
  cbind(scIntegrationMetrics::meta_data) \%>\% 
  sample_frac(1L, FALSE) \%>\% 
  gather(key, val, label1, label2) \%>\% 
  ggplot(aes(X1, X2, color = val)) +
    geom_point(shape = 21) + 
    facet_wrap(~key)
  
## Now to compute and plot the LISI values for each label. 
lisi_res <- compute_lisi(scIntegrationMetrics::X, scIntegrationMetrics::meta_data, c('label1', 'label2'))
head(lisi_res)

  cbind(lisi_res) \%>\% 
  sample_frac(1L, FALSE) \%>\% 
  gather(key, lisi_value, label1, label2) \%>\% 
  ggplot(aes(X1, X2, color = lisi_value)) +
    geom_point(shape = 21) + 
    facet_wrap(~key)

}
