% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_mean_singleLevel}
\alias{compute_mean_singleLevel}
\title{Single-level means}
\usage{
compute_mean_singleLevel(res, meta_data, label_colnames, level)
}
\arguments{
\item{meta_data}{A data frame with one row per cell.}

\item{label_colnames}{Which variables to compute averages for.}

\item{level}{Which level to consider}

\item{X}{A matrix with cells (rows) and features (columns).}
}
\value{
A vector of means
}
\description{
Single-level means
}
