% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getIntegrationMetrics}
\alias{getIntegrationMetrics}
\title{Compute multiple integration metrics from a Seurat object}
\usage{
getIntegrationMetrics(
  object,
  metrics = NULL,
  meta.label,
  meta.batch,
  lisi_perplexity = 30,
  method.reduction = "pca",
  metricsLabels = NULL
)
}
\arguments{
\item{object}{A seurat object with dimensionality reduction and meta data}

\item{metrics}{one or more of 'batch_LISI', 'batch_nLISI', 'batch_nLISI_means', 'batch_nLISI_perCellType', 'batch_nLISI_perCellType_means','celltype_nLISI', 'celltype_nLISI_means', 'celltype_ASW', 'celltype_ASW_means', or leave 'NULL' to calculate them all}

\item{meta.label}{Which meta data column contains cluster/celltype labels}

\item{meta.batch}{Which meta data column contains batch}

\item{method.reduction}{reduction method to consider, eg 'pca'}

\item{metricsLabels}{which cluster/celltype labels to consider to summarize metrics, by default use all}
}
\value{
A list of mean values for each metric
}
\description{
Compute multiple integration metrics from a Seurat object
}
