% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegrationMetrics.R
\docType{package}
\name{scIntegrationMetrics-package}
\alias{scIntegrationMetrics}
\alias{scIntegrationMetrics-package}
\title{scIntegrationMetrics}
\description{
A collection of metrics to evaluate integration of scRNA-seq datasets.
}
\author{
\strong{Maintainer}: Massimo Andreatta \email{massimo.andreatta@unil.ch} (\href{https://orcid.org/0000-0002-8036-2647}{ORCID})

Authors:
\itemize{
  \item Josep Garnica \email{josep.garnicacaparros@unil.ch} (\href{https://orcid.org/0000-0001-9493-1321}{ORCID})
  \item Santiago Carmona \email{santiago.carmona@unil.ch} (\href{https://orcid.org/0000-0002-2495-0671}{ORCID})
}

}
\keyword{internal}
