% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_lisi_splitBy}
\alias{compute_lisi_splitBy}
\title{Compute Local Inverse Simpson's Index (LISI) split by groups}
\usage{
compute_lisi_splitBy(
  X,
  meta_data,
  label_colnames,
  split_by_colname,
  metricsLabels = NULL,
  normalize = TRUE,
  ...
)
}
\arguments{
\item{X}{A matrix with cells (rows) and features (columns).}

\item{meta_data}{A data frame with one row per cell.}

\item{label_colnames}{Which variables to compute averages for.}

\item{split_by_colname}{Which variable levels use to split data}

\item{metricsLabels}{Which cell types to evaluate for LISI calculation}

\item{normalize}{Normalize LISI between 0 and 1}
}
\value{
A list of data frames of LISI values (one per split_by_colname level). Each row is a cell and each
column is a different label variable.
}
\description{
Compute Local Inverse Simpson's Index (LISI) split by groups
}
